Implemented recommended changes:  - page[totals] is now a valueless parameter. The presence of this parameter will cause page totals to be generated.  - Paginate annotation supported on the entity class including support for countable, defaultLimit, maxLimit  - tests
Checking permissions when generating totals remains to be addressed. It seems to me that this introduces potentially large inefficiency as the full result set would need to be loaded to check permissions. Maybe it should be accepted that a limitation of totals generation is that permissions will not be checked - similar to cascade delete.