Avoid checking share permission for redundant add
We have been calling addRelation to parent entity for updates to child
entities in graphql. This checks the share permission for the child
entity and may fail in some cases.
This PR avoids checking the share permission in case when the entity is
already a part of the relation with the parent.